+++ client/src/app/models/data-source.ts
@@ -22,8 +22,7 @@ export enum DataSourceConnectionStatus {
   Reconnecting = 'reconnecting',
   ConnectionError = 'failedtoconnect',
   Unavailable = 'unavailable', // in case of status can't be loaded
-  TermsAndConditionsNotAccepted = 'termsandconditionsnotaccepted',
-  NoLicense = 'nolicense'
+  TermsAndConditionsNotAccepted = 'termsandconditionsnotaccepted'
 }
 
 export class DataSourceConnection {