+++ src/modules/Backend/AuthManager/index.ts
@@ -91,17 +91,19 @@ export class AuthManager {
   }
 
   /**
-   * @param  {{withPasswordChangeDetection:boolean}} options
+   * @param  {{withPasswordChangeDetection:boolean,withBooleanResponse:(boolean|undefined)}} options
    * @returns {Function} callback
    */
   verifyJWTAuth({
-    withPasswordChangeDetection
+    withPasswordChangeDetection,
+    withBooleanResponse
   }: {
     withPasswordChangeDetection: boolean;
+    withBooleanResponse?: boolean;
   }) {
     const logPrefix = `${AuthManager.className}::verifyJWTAuth`;
 
-    return (request: Request, response: Response, next: NextFunction) => {
+    return (request: Request, response: Response, next?: NextFunction) => {
       const header = request.headers['authorization'];
 
       if (!header) {
@@ -130,13 +132,19 @@ export class AuthManager {
         }
 
         request.user = user;
-
-        next();
+        if (withBooleanResponse) {
+          return true;
+        } else {
+          next();
+        }
       } catch (err) {
         winston.error(
           `${logPrefix} jwt vaidation failed. ${JSON.stringify(err)}`
         );
         response.status(401).json({ message: 'Unauthorized!' });
+        if (withBooleanResponse) {
+          return false;
+        }
       }
     };
   }