+++ src/modules/Southbound/DataSources/DataSource.ts
@@ -19,7 +19,6 @@ import {
 import Timeout = NodeJS.Timeout;
 import { SynchronousIntervalScheduler } from '../../SyncScheduler';
 import winston from 'winston';
-import { LicenseChecker } from '../../LicenseChecker';
 
 type DataPointReadErrorSummary = {
   error: string;
@@ -47,7 +46,6 @@ export abstract class DataSource extends EventEmitter {
   protected processedDataPointCount = 0;
   protected dataPointReadErrors: DataPointReadErrorSummary[] = [];
   protected readCycleCount = 0;
-  protected isLicensed = false;
 
   /**
    * Create a new instance & initialize the sync scheduler
@@ -59,7 +57,6 @@ export abstract class DataSource extends EventEmitter {
     this.scheduler = SynchronousIntervalScheduler.getInstance();
     this.protocol = params.config.protocol;
     this.termsAndConditionsAccepted = params.termsAndConditionsAccepted;
-    this.isLicensed = params.isLicensed;
   }
 
   /**
@@ -201,13 +198,12 @@ export abstract class DataSource extends EventEmitter {
   protected onDataPointMeasurement = (measurements: IMeasurement[]): void => {
     const logPrefix = `${this.name}::onDataPointMeasurement`;
 
-    const { name, protocol } = this.config;
+    const { protocol } = this.config;
 
     try {
       this.submitMeasurement(
         measurements.map((measurement) => ({
           dataSource: {
-            name,
             protocol
           },
           measurement
@@ -226,10 +222,9 @@ export abstract class DataSource extends EventEmitter {
   protected onDataPointLifecycle = (
     lifecycleEvent: IBaseLifecycleEvent
   ): void => {
-    const { name, protocol } = this.config;
+    const { protocol } = this.config;
     const DPLifecycleEvent: IDataSourceDataPointLifecycleEvent = {
       dataSource: {
-        name,
         protocol
       },
       ...lifecycleEvent