+++ src/modules/System/index.ts
@@ -1,7 +1,7 @@
 import { promises as fs, readFileSync, existsSync } from 'fs';
+import fetch from 'node-fetch';
 import { promisify } from 'util';
 import winston from 'winston';
-import SshService from '../SshService';
 const child_process = require('child_process');
 const exec = promisify(child_process.exec);
 
@@ -18,8 +18,8 @@ export class System {
   ): Promise<string | null> {
     let address;
     enum interfaceMapping {
-      'eth0' = 'eno1',
-      'eth1' = 'eno2'
+      'eth0' = 'enoX1',
+      'eth1' = 'enoX1'
     }
     try {
       address = await fs.readFile(
@@ -57,45 +57,46 @@ export class System {
    * @async
    * @returns {Promise<string>}
    */
-  public async getHostname() {
-    const macAddress = (await this.readMacAddress('eth0')) || '000000000000';
-    const formattedMacAddress = macAddress.split(':').join('').toUpperCase();
-    return `DM${formattedMacAddress}`;
+  public async getHostname(): Promise<string> {
+    const defaultHostname = 'dm000000000000';
+    return process.env.IOTEDGE_GATEWAYHOSTNAME || defaultHostname;
   }
 
   /**
-   * Restarts device
+   * Reads CELOS version
    */
-  public async restartDevice() {
-    const logPrefix = `${System.className}::restartDevice`;
+  public async readOsVersion() {
     try {
-      winston.info(`${logPrefix} restarting device`);
-      await SshService.sendCommand('reboot', true);
-    } catch (err) {}
+      const PROXY_HOST =
+        process.env.NODE_ENV === 'development'
+          ? 'http://localhost'
+          : 'http://172.17.0.1';
+      const PROXY_PORT = 1884;
+      const PATH_PREFIX = '/api/v1';
 
-    process.exit(0);
-  }
-
-  /**
-   * Reads mdc flex OS version. Located
-   */
-  public readOsVersion() {
-    try {
-      if (existsSync('/etc/mdcflex-os-release')) {
-        const fileContent = readFileSync(
-          '/etc/mdcflex-os-release',
-          'utf-8'
-        ).trim();
-        const versionLine =
-          fileContent.split('\n').find((line) => line.startsWith('VERSION=')) ||
-          'unknown';
-        return versionLine.replace('VERSION=', '');
+      let response = await fetch(
+        `${PROXY_HOST}:${PROXY_PORT}${PATH_PREFIX}/system/versions`,
+        {
+          method: 'GET',
+          headers: {
+            Accept: 'application/json',
+            'Content-Type': 'application/json'
+          }
+        }
+      );
+      if (response.ok) {
+        const versions = await response?.json();
+        const osVersion = versions?.find((v) => v.Name === 'COS')?.Version;
+        return osVersion || 'unknown';
+      } else {
+        winston.error('OS Version reading response is not OK');
+        winston.error(JSON.stringify(response));
+        return 'unknown';
       }
     } catch (e) {
-      console.log('Error reading OS version!');
-      console.log(e);
+      winston.error('Error reading OS version!');
+      winston.error(e.message);
+      return 'unknown';
     }
-
-    return 'unknown';
   }
 }